backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
nba_api>=1.4.1
cachetools>=5.3.3
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pytest>=8.2.0
pytest-asyncio>=0.23.7
httpx>=0.27.0
